post delete deadlock

+++ modules/containers/custom_field.php
@@ -590,14 +590,16 @@ class blcPostMetaManager extends blcContainerManager {
    */
 	function post_deleted($post_id){
 		//Get the associated container object
-		$container = blcContainerHelper::get_container( array($this->container_type, intval($post_id)) );
-		//Delete it
-		$container->delete();
-		//Clean up any dangling links
-		blc_cleanup_links();
-	}
-	
-  /**
+        $container = blcContainerHelper::get_container([$this->container_type, intval($post_id)]);
+        if ($container != null) {
+            //Delete it
+            $container->delete();
+            //Clean up any dangling links
+            blc_cleanup_links();
+        }
+    }
+
+    /**
    * When a post is restored, mark all of its custom fields as unparsed.
    * Called via the 'untrashed_post' action.
    *
